starting implementation of map_block functions + directory hierarchy refactor
