+++ seapodym_lmtl_python/configuration/no_transport/parameter_forcing.py
@@ -11,7 +11,7 @@ import pandas as pd
 import xarray as xr
 from attrs import converters, field, frozen
 
-from seapodym_lmtl_python.cf_data.units import check_units
+from seapodym_lmtl_python.standard.units import check_units
 from seapodym_lmtl_python.logging.custom_logger import logger
 
 DECIMALS = 5  # ie. 1e-5 degrees which is equivalent to ~1m at the equator
