+++ seapodym_lmtl_python/post_production/no_transport/biomass.py
@@ -1,0 +1,1 @@
+"""Wrap the biomass computation to follow the no-transport model."""
