+++ seapodym_lmtl_python/production/no_transport/production.py
@@ -1,0 +1,1 @@
+"""Wrap the biomass computation to follow the no-transport model."""
