+++ test/configuration/no_transport/test_configuration.py
@@ -2,17 +2,17 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from seapodym_lmtl_python.cf_data import coordinates
 from seapodym_lmtl_python.configuration.no_transport import parameter_functional_group
 from seapodym_lmtl_python.configuration.no_transport.configuration import NoTransportConfiguration
-from seapodym_lmtl_python.configuration.no_transport.labels import StandardUnitsLabels
-from seapodym_lmtl_python.configuration.no_transport.parameter_environment import EnvironmentParameter
-from seapodym_lmtl_python.configuration.no_transport.parameter_forcing import ForcingUnit
-from seapodym_lmtl_python.configuration.no_transport.parameters import (
+from seapodym_lmtl_python.configuration.no_transport.parameter import (
     ForcingParameters,
     FunctionalGroups,
     NoTransportParameters,
 )
+from seapodym_lmtl_python.configuration.no_transport.parameter_environment import EnvironmentParameter
+from seapodym_lmtl_python.configuration.no_transport.parameter_forcing import ForcingUnit
+from seapodym_lmtl_python.standard import coordinates
+from seapodym_lmtl_python.standard.units import StandardUnitsLabels
 
 
 @pytest.fixture()
