+++ test/pre_production/core/test_day_length.py
@@ -3,8 +3,8 @@ import pint
 import pytest
 import xarray as xr
 
-from seapodym_lmtl_python.cf_data import coordinates
 from seapodym_lmtl_python.pre_production.core.day_length import day_length_forsythe
+from seapodym_lmtl_python.standard import coordinates
 
 
 @pytest.fixture()
